frontend/requirements.txt
python-chess
pygame
numpy
tqdm
